+++ lib/auth_trail/engine.rb
@@ -1,7 +1,0 @@
-require "active_support"
-require "rails/engine"
-
-module AuthTrail
-  class Engine < Rails::Engine
-  end
-end