+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.5.0 (unreleased)
 
+- Made Active Job optional
 - Dropped support for Ruby < 3 and Rails < 6.1
 
 ## 0.4.3 (2022-06-12)